+++ R/TPRdim.R
@@ -36,7 +36,7 @@
 #' TPRdim(x, y, maxdim = 5)
 #'
 #' ## Use dataset square. (time-consuming)
-#' \dontrun{
+#' \donttest{
 #' data("square")
 #' x <- square$x
 #' y <- square$y