+++ R/TRRdim.R
@@ -36,8 +36,8 @@
 #'
 #' TRRdim(x, y) # The estimated envelope dimensions are the same as u.
 #'
-#' ## Use dataset bat, but it is time-consuming
-#' \dontrun{
+#' ## Use dataset bat. (time-consuming)
+#' \donttest{
 #' data("bat")
 #' x <- bat$x
 #' y <- bat$y